Finished organizational refactor
