--- src/containers/app/routes/static/homepage/_timeline.scss
@@ -1,251 +1,0 @@
-$timeline-item-height: 20px;
-$github-project-size: 270px;
-
-#timeline {
-  padding: $large-space 0px;
-  @include desktop-up {
-    padding: $xlarge-space + $medium-space 0px;
-  }
-  #project-timeline {
-    .events {
-      list-style-type: none;
-      padding: 0px;
-      @include desktop-up {
-        display: flex;
-        justify-content: space-between;
-      }
-      .event {
-        display: flex;
-        @include desktop-up {
-          display: block;
-        }
-        &.future {
-          opacity: 0.5;
-        }
-        &.present .marker {
-          background: $primary-color;
-          box-shadow: 0px 0px 0px 2px $primary-color;
-        }
-        .marker {
-          display: block;
-          height: $timeline-item-height;
-          width: $timeline-item-height;
-          border-radius: $round-radius;
-          background: $success-color;
-          border: 2px solid $white;
-          box-shadow: 0px 0px 0px 2px $success-color;
-        }
-        .content {
-          margin-left: $medium-space;
-          margin-bottom: $small-space;
-          @include desktop-up {
-            margin: 0px;
-            margin-top: $medium-space;
-          }
-          .title {
-            text-transform: none;
-            letter-spacing: normal;
-            margin: 0px;
-            @include tablet-only {
-              margin-top: -4px;
-            }
-          }
-          .date {
-            @include capped-text(1px);
-            font-size: 0.85rem;
-            color: $medium-gray;
-            margin-top: 5px;
-          }
-        }
-      }
-    }
-    @include desktop-up() {
-      position: relative;
-      width: 100%;
-      margin-top: $large-space;
-      &:before {
-        position: absolute;
-        top: 25px;
-        left: 0px;
-        height: 2px;
-        width: 100%;
-        display: block;
-        z-index: -1;
-        background: $light-gray;
-        content: '';
-      }
-    }
-  }
-  #github-projects {
-    .projects {
-      display: flex;
-      justify-content: space-around;
-      flex-wrap: wrap;
-      margin: $medium-space auto;
-      @include desktop-up {
-        margin: $large-space auto;
-      }
-      .project {
-        display: flex;
-        flex-direction: column;
-        justify-content: center;
-        align-items: center;
-        margin-bottom: $medium-space;
-        background: $white;
-        padding: $medium-space;
-        text-align: center;
-        border-radius: $extra-border-radius;
-        box-shadow: 0px $small-space $large-space rgba($black, 0.1);
-        cursor: pointer;
-        width: 100%;
-        height: $github-project-size * 0.35;
-        transition: opacity $anim-time $anim-func,
-          transform $anim-time $anim-func;
-        opacity: 0.35;
-        @include tablet-up {
-          flex-basis: 45%;
-          width: $github-project-size;
-          height: $github-project-size;
-        }
-        @include widescreen-up {
-          flex-basis: auto;
-          width: $github-project-size;
-          height: $github-project-size;
-        }
-        &:hover {
-          transform: scale(1.05);
-        }
-        &.current {
-          opacity: 1;
-          &:hover {
-            transform: none;
-          }
-        }
-        .repo-icon {
-          height: 48px;
-          display: none;
-          @include tablet-up {
-            display: block;
-          }
-        }
-        .heading {
-          margin-top: $small-space;
-          @include tablet-up {
-            margin-top: $medium-space;
-          }
-        }
-        .description {
-          margin-top: 0px;
-          font-size: 0.9rem;
-          color: $medium-gray;
-        }
-      }
-    }
-    .contributors {
-      list-style-type: none;
-      padding: 0px;
-      .contributor {
-        a {
-          display: flex;
-          align-items: center;
-          margin-bottom: $medium-space;
-          &:hover .login {
-            color: $primary-color;
-          }
-          .avatar {
-            height: 48px;
-            width: 48px;
-            display: block;
-            border-radius: $round-radius;
-            background-size: cover;
-            background-position: 50% 50%;
-          }
-          .login {
-            font-weight: $bold-weight;
-            margin: 0px;
-            margin-left: $small-space;
-            color: $black;
-            transition: color $anim-time $anim-func;
-          }
-        }
-      }
-    }
-    .issues {
-      list-style-type: none;
-      padding: 0px;
-      .no-issues span {
-        color: $medium-gray;
-      }
-      .issue {
-        display: flex;
-        margin-bottom: $medium-space;
-        .fa-github {
-          font-size: 1.5rem;
-          height: 24px;
-          display: block;
-          margin-right: $medium-space;
-          opacity: 0.25;
-        }
-        a {
-          display: block;
-          &:hover .title {
-            color: $primary-color;
-          }
-          .title {
-            margin: 0px;
-            font-size: 1.1rem;
-            font-weight: $bold-weight;
-            color: $black;
-            transition: color $anim-time $anim-func;
-          }
-          .meta {
-            margin: 0px;
-            font-size: 0.9rem;
-            color: $medium-gray;
-          }
-        }
-      }
-    }
-  }
-  .cta-container {
-    display: flex;
-    justify-content: center;
-    margin-top: $medium-space;
-    @include desktop-up {
-      margin-top: $large-space;
-    }
-    .button {
-      position: relative;
-      display: flex;
-      align-items: center;
-      &:before,
-      &:after {
-        position: absolute;
-        top: 50%;
-        width: 100%;
-        height: 1px;
-        content: '';
-      }
-      &:before {
-        left: -260px;
-        background: linear-gradient(
-          to right,
-          rgba($light-gray, 0),
-          rgba($light-gray, 1)
-        );
-      }
-      &:after {
-        right: -260px;
-        background: linear-gradient(
-          to left,
-          rgba($light-gray, 0),
-          rgba($light-gray, 1)
-        );
-      }
-      i {
-        margin-right: $small-space;
-        font-size: 1.1rem;
-      }
-    }
-  }
-}
