+++ src/containers/app/routes/static/homepage/footer/footer.scss
@@ -1,3 +1,5 @@
+@import 'variables';
+
 #footer {
   background: linear-gradient(to bottom, $dark-gray, darken($dark-gray, 7%));
   padding-top: $medium-space;
