--- src/containers/app/routes/static/homepage/homepage.scss
@@ -1,7 +1,0 @@
-@import 'variables';
-
-#homepage {
-  @import 'process';
-  @import 'timeline';
-  @import 'footer';
-}
