+++ src/containers/app/routes/static/homepage/pillars/pillars.scss
@@ -14,6 +14,7 @@
       @include desktop-up() {
         display: flex;
         flex-direction: column;
+        height: 100%;
       }
       .icon {
         display: block;
