+++ src/containers/app/routes/static/homepage/process/_graph.scss
@@ -1,0 +1,203 @@
+$process-height: 360px;
+$process-node: 80px;
+$process-node-icon: 60%;
+
+@keyframes distributingAnimation {
+  0% {
+    background: $black;
+  }
+  50% {
+    background: $warning-color;
+  }
+  100% {
+    background: $black;
+  }
+}
+
+@keyframes dottedAnimation {
+  0% {
+    opacity: 0;
+  }
+  50% {
+    opacity: 1;
+  }
+  100% {
+    opacity: 0;
+  }
+}
+
+.graph {
+  display: none;
+  @include desktop-up {
+    display: flex;
+    width: 100%;
+    justify-content: space-between;
+    height: $process-height * 0.8;
+    @include widescreen-up {
+      height: $process-height;
+    }
+    .items {
+      list-style-type: none;
+      padding: 0px;
+      margin: 0px;
+      display: flex;
+      flex-direction: column;
+      justify-content: space-between;
+      align-items: center;
+      .repo-icon {
+        transition: all $anim-time $anim-func;
+      }
+      &.disabled {
+        .repo-icon {
+          filter: grayscale(100%);
+          opacity: 0.25;
+        }
+      }
+      &.data-scientists {
+        justify-content: center;
+        .repo-icon {
+          filter: grayscale(100%);
+        }
+        &.finished {
+          .repo-icon {
+            filter: grayscale(0%);
+          }
+        }
+      }
+      &.sonar {
+        .big,
+        .small {
+          transition: background $anim-time $anim-func;
+        }
+        .big {
+          width: $process-node / 2;
+          height: $process-node / 2;
+          opacity: 0.66;
+        }
+        .small {
+          width: $process-node / 4;
+          height: $process-node / 4;
+          opacity: 0.33;
+        }
+        &.distributing {
+          .big,
+          .small {
+            animation: distributingAnimation 1s $anim-func infinite;
+          }
+        }
+        &.distributed {
+          .big,
+          .small {
+            background: $warning-color;
+          }
+        }
+      }
+      &.mines {
+        &.money {
+          .repo-icon {
+            filter: hue-rotate(100deg);
+          }
+        }
+      }
+      .item {
+        position: relative;
+        display: block;
+        width: $process-node;
+        height: $process-node;
+        background: $black;
+        border-radius: $border-radius;
+        img {
+          position: absolute;
+          top: 50%;
+          left: 50%;
+          width: $process-node-icon;
+          height: $process-node-icon;
+          margin-left: -0.5 * $process-node-icon;
+          margin-top: -0.5 * $process-node-icon;
+          transition: opacity $anim-time $anim-func;
+        }
+      }
+    }
+    .lines {
+      list-style-type: none;
+      padding: 0px;
+      margin: 0px;
+      display: flex;
+      flex-direction: column;
+      justify-content: center;
+      align-items: center;
+      flex-basis: 100%;
+      &.disabled {
+        .line:after {
+          opacity: 0;
+          animation: none;
+        }
+      }
+      &.first {
+        li {
+          &:after {
+            background: linear-gradient(to right, $light-gray, $warning-color);
+            clip-path: url(#circles-clip);
+          }
+        }
+        &.returned li:after {
+          background: linear-gradient(to right, $success-color, $warning-color);
+        }
+      }
+      &.second {
+        li {
+          transition: background $anim-time $anim-func;
+          &:first-of-type {
+            transform-origin: 0% 0%;
+            transform: rotate(-35deg) scaleX(1.2);
+            @include widescreen-up {
+              transform: rotate(-28deg) scaleX(1.125);
+            }
+          }
+          &:last-of-type {
+            transform-origin: 0% 0%;
+            transform: translateX(9px) rotate(35deg) scaleX(1.2);
+            @include widescreen-up {
+              transform: translateX(9px) rotate(28deg) scaleX(1.125);
+            }
+          }
+          &:after {
+            background: linear-gradient(to right, $warning-color, $alert-color);
+            clip-path: url(#circles-clip);
+          }
+        }
+        &.rewarded li:after {
+          background: linear-gradient(to right, $warning-color, $success-color);
+          clip-path: url(#circles-clip);
+        }
+      }
+      li {
+        display: block;
+        width: 100%;
+        height: 20px;
+        position: relative;
+        &:before {
+          position: absolute;
+          top: 50%;
+          left: 0px;
+          width: 100%;
+          height: 2px;
+          border: 2px solid rgba($white, 0.1);
+          overflow: hidden;
+          margin-top: -2px;
+          content: '';
+        }
+        &:after {
+          position: absolute;
+          top: 0px;
+          left: 0px;
+          width: 100%;
+          height: 100%;
+          content: '';
+          transition: opacity $anim-time $anim-func;
+          animation: dottedAnimation 1s $anim-func infinite;
+        }
+      }
+    }
+  }
+}
