+++ src/containers/app/routes/static/homepage/process/_step-selector.scss
@@ -1,0 +1,63 @@
+$process-timer-size: 6px;
+
+@keyframes progress {
+  from {
+    width: 0%;
+  }
+
+  to {
+    width: 100%;
+  }
+}
+
+.step-selector {
+  padding: $medium-space 0px 0px 0px;
+  @include desktop-up {
+    padding: $xlarge-space + $medium-space 0px 0px 0px;
+  }
+  .step-progress {
+    position: relative;
+    height: $process-timer-size;
+    width: 0%;
+    background: $primary-color;
+    transition: height $anim-time $anim-func;
+    &.stopped {
+      height: 0px;
+      animation-play-state: paused;
+    }
+  }
+  .step-selector-list {
+    padding-left: 0px;
+    list-style-type: none;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    margin: 0px;
+    @include tablet-up {
+      flex-direction: row;
+      justify-content: space-between;
+    }
+    .step-item {
+      @include capped-text(2px);
+      text-align: center;
+      width: 100%;
+      padding: $small-space 0px;
+      flex-grow: 1;
+      background: $dark-gray;
+      color: $white;
+      cursor: pointer;
+      opacity: 0.5;
+      transition: opacity $anim-time $anim-func;
+      @include desktop-up {
+        font-size: 1.1rem;
+        padding: $medium-space 0px;
+      }
+      &:hover {
+        opacity: 1;
+      }
+      &.active {
+        opacity: 1;
+      }
+    }
+  }
+}
