+++ src/containers/app/routes/static/homepage/process/process.scss
@@ -1,0 +1,21 @@
+@import 'variables';
+
+#process {
+  background: linear-gradient(to bottom, $dark-gray, darken($dark-gray, 7%));
+  padding: $large-space 0px 0px 0px;
+  @include desktop-up {
+    padding: $xlarge-space + $medium-space 0px 0px 0px;
+  }
+  h3 {
+    color: $white;
+  }
+  @import 'step-selector';
+  .step-content {
+    margin-top: $medium-space;
+    @include desktop-up {
+      margin-top: $large-space;
+    }
+    @import 'graph';
+    @import 'info';
+  }
+}
