+++ src/containers/app/routes/static/homepage/status/_projects.scss
@@ -1,0 +1,130 @@
+#github-projects {
+  .projects {
+    display: flex;
+    justify-content: space-around;
+    flex-wrap: wrap;
+    margin: $medium-space auto;
+    @include desktop-up {
+      margin: $large-space auto;
+    }
+    .project {
+      display: flex;
+      flex-direction: column;
+      justify-content: center;
+      align-items: center;
+      margin-bottom: $medium-space;
+      background: $white;
+      padding: $medium-space;
+      text-align: center;
+      border-radius: $extra-border-radius;
+      box-shadow: 0px $small-space $large-space rgba($black, 0.1);
+      cursor: pointer;
+      width: 100%;
+      height: $github-project-size * 0.35;
+      transition: opacity $anim-time $anim-func, transform $anim-time $anim-func;
+      opacity: 0.35;
+      @include tablet-up {
+        flex-basis: 45%;
+        width: $github-project-size;
+        height: $github-project-size;
+      }
+      @include widescreen-up {
+        flex-basis: auto;
+        width: $github-project-size;
+        height: $github-project-size;
+      }
+      &:hover {
+        transform: scale(1.05);
+      }
+      &.current {
+        opacity: 1;
+        &:hover {
+          transform: none;
+        }
+      }
+      .repo-icon {
+        height: 48px;
+        display: none;
+        @include tablet-up {
+          display: block;
+        }
+      }
+      .heading {
+        margin-top: $small-space;
+        @include tablet-up {
+          margin-top: $medium-space;
+        }
+      }
+      .description {
+        margin-top: 0px;
+        font-size: 0.9rem;
+        color: $medium-gray;
+      }
+    }
+  }
+  .contributors {
+    list-style-type: none;
+    padding: 0px;
+    .contributor {
+      a {
+        display: flex;
+        align-items: center;
+        margin-bottom: $medium-space;
+        &:hover .login {
+          color: $primary-color;
+        }
+        .avatar {
+          height: 48px;
+          width: 48px;
+          display: block;
+          border-radius: $round-radius;
+          background-size: cover;
+          background-position: 50% 50%;
+        }
+        .login {
+          font-weight: $bold-weight;
+          margin: 0px;
+          margin-left: $small-space;
+          color: $black;
+          transition: color $anim-time $anim-func;
+        }
+      }
+    }
+  }
+  .issues {
+    list-style-type: none;
+    padding: 0px;
+    .no-issues span {
+      color: $medium-gray;
+    }
+    .issue {
+      display: flex;
+      margin-bottom: $medium-space;
+      .fa-github {
+        font-size: 1.5rem;
+        height: 24px;
+        display: block;
+        margin-right: $medium-space;
+        opacity: 0.25;
+      }
+      a {
+        display: block;
+        &:hover .title {
+          color: $primary-color;
+        }
+        .title {
+          margin: 0px;
+          font-size: 1.1rem;
+          font-weight: $bold-weight;
+          color: $black;
+          transition: color $anim-time $anim-func;
+        }
+        .meta {
+          margin: 0px;
+          font-size: 0.9rem;
+          color: $medium-gray;
+        }
+      }
+    }
+  }
+}
